+++ NEWS.md
@@ -1,5 +1,7 @@
-ctmm 1.1.1 (2022-11-04)
+ctmm 1.1.1 (2022-11-15)
 ================
+
+  * new diagonal population covariance models to select in mean.ctmm()
 
   * new argument sqrt in distance()